Change restplus app initialization process
